feat: add component statistics retrieval for the last 30 days and update related entities

+++ src/main/java/org/project/backend_pcbuild/pcComponents/dto/ComponentsAmountPc.java
@@ -2,5 +2,7 @@ package org.project.backend_pcbuild.pcComponents.dto;
 
 import org.project.backend_pcbuild.pcComponents.model.ComponentType;
 
-public record ComponentsAmountPc(String componentType, String model, Long amount) {
+import java.math.BigInteger;
+
+public record ComponentsAmountPc(String componentType, String model, Integer amount) {
 }
